requirements.txt
# FastAPI + Server
fastapi==0.115.6
uvicorn[standard]==0.34.0
python-multipart==0.0.20
websockets==14.2

# MySQL + ORM
sqlalchemy[asyncio]==2.0.36
pymysql==1.1.1
cryptography==44.0.0
alembic==1.14.1
aiomysql==0.2.0

# MinIO
minio==7.2.12

# Redis
redis[hiredis]==5.2.1

# Celery
celery[redis]==5.4.0

# DuckDB
duckdb==1.1.3

# Auth
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
bcrypt==4.2.1

# Scheduling
apscheduler==3.10.4

# Data processing
pyarrow==18.1.0
pandas==2.2.3
openpyxl==3.1.5
xlsxwriter==3.2.0
lxml==5.3.0

# HTTP
httpx==0.28.1

# PDF
pdfplumber==0.11.4

# CLI
typer[all]==0.15.1

# Validation / Security
pydantic==2.10.4
pydantic-settings==2.7.1
python-dotenv==1.0.1
bleach==6.2.0
python-magic==0.4.27
slowapi==0.1.9

# Email
aiosmtplib==3.0.2

# XML-RPC (stdlib, no pip needed)
# FTP (ftplib - stdlib)

# Testing
pytest==8.3.4
pytest-asyncio==0.25.0
httpx==0.28.1
pytest-cov==6.0.0

# Utilities
python-dateutil==2.9.0.post0
orjson==3.10.12
aiofiles==24.1.0
Jinja2==3.1.5
